Moved streamer code to streamer
